+++ socialhome/federate/admin.py
@@ -6,6 +6,6 @@ from socialhome.federate.models import Payload
 
 @admin.register(Payload)
 class PayloadAdmin(ModelAdmin):
-    list_display = ('id', 'entities_found', 'protocol', 'method', 'url', 'created')
-    list_filter = ('entities_found', 'protocol', 'method', 'created')
+    list_display = ('id', 'entities_found', 'protocol', 'direction', 'method', 'url', 'created')
+    list_filter = ('entities_found', 'direction', 'protocol', 'method', 'created')
     search_fields = ('body', 'sender', 'url')